backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.1
tensorflow==2.18.0
pillow==10.0.0
numpy==1.26.5
